# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CodeForge Academy API",
    "version": "v1"
  },
  "paths": {
...
    "/modules/{moduleId}/assignments": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponseDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{id}": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponseDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/modules/{moduleId}/assignments/reorder": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderAssignmentsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderAssignmentsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderAssignmentsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/assignments/{assignmentId}/test-cases": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTestCaseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTestCaseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddTestCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCaseResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/test-cases/{id}": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestCaseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestCaseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCaseResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCaseResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignmentId}/test-cases/reorder": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTestCasesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTestCasesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTestCasesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/assignments/{id}/submission": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentForSubmissionDto"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{id}/submissions": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssignmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssignmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResultDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentSubmissionsDto"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submissionId}/grade": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "submissionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeSubmissionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeSubmissionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GradeSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{id}/my-submissions": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submissionId}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "submissionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddTestCaseRequest": {
        "required": [
          "expectedOutput",
          "input",
          "isHidden",
          "points"
        ],
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "expectedOutput": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "AssignmentDetailDto": {
        "required": [
          "description",
          "id",
          "isPractice",
          "moduleId",
          "orderIndex",
          "testCases",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "isPractice": {
            "type": "boolean"
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "passScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "testCases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCaseDto"
            }
          }
        },
        "additionalProperties": false
      },
      "AssignmentDto": {
        "required": [
          "createdAt",
          "description",
          "id",
          "isPractice",
          "moduleId",
          "orderIndex",
          "testCaseCount",
          "title",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "isPractice": {
            "type": "boolean"
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "passScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "testCaseCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AssignmentForSubmissionDto": {
        "required": [
          "attemptsUsed",
          "description",
          "id",
          "sampleTestCases",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "attemptsUsed": {
            "type": "integer",
            "format": "int32"
          },
          "sampleTestCases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionTestCaseDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "AssignmentResponseDto": {
        "required": [
          "assignmentId",
          "message"
        ],
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AssignmentSubmissionsDto": {
        "required": [
          "assignmentId",
          "assignmentTitle",
          "submissions"
        ],
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "assignmentTitle": {
            "type": "string"
          },
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentSubmissionDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "CreateAssignmentRequest": {
        "required": [
          "description",
          "isPractice",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPractice": {
            "type": "boolean"
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "passScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GradeSubmissionRequest": {
        "required": [
          "manualScore"
        ],
        "type": "object",
        "properties": {
          "manualScore": {
            "type": "integer",
            "format": "int32"
          },
          "manualFeedback": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReorderAssignmentsRequest": {
        "required": [
          "assignmentOrders"
        ],
        "type": "object",
        "properties": {
          "assignmentOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssignmentOrderDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ReorderTestCasesRequest": {
        "required": [
          "testCaseOrders"
        ],
        "type": "object",
        "properties": {
          "testCaseOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCaseOrderDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "SubmissionResultDto": {
        "required": [
          "attemptNumber",
          "autoGradingStatus",
          "isLate",
          "submissionId",
          "submittedAt",
          "testResults"
        ],
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "string",
            "format": "uuid"
          },
          "attemptNumber": {
            "type": "integer",
            "format": "int32"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isLate": {
            "type": "boolean"
          },
          "autoScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "autoGradingStatus": {
            "type": "string"
          },
          "manualScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "manualFeedback": {
            "type": "string",
            "nullable": true
          },
          "finalScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "testResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestResultDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "SubmitAssignmentRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "TestCaseResponseDto": {
        "required": [
          "message",
          "testCaseId"
        ],
        "type": "object",
        "properties": {
          "testCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAssignmentRequest": {
        "required": [
          "description",
          "isPractice",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPractice": {
            "type": "boolean"
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "passScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTestCaseRequest": {
        "required": [
          "expectedOutput",
          "input",
          "isHidden",
          "points"
        ],
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "expectedOutput": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}